+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ENHANCEMENTS:
 
 BUG FIXES:
+* resource/`junos_security_nat_source`: fix order issue on `from.0.value` and `to.0.value` lists (Fixes #243)
 
 ## 1.17.0 (June 18, 2021)
 FEATURES: